chore(web): P9 — dead UI dependencies pruned; line-chart-6 retired to its types
package.json: removed 14 @radix-ui/*, radix-ui, react-aria-components,
tailwind-variants, @intentui/icons, vaul (all grep-verified zero
imports; ui/drawer is Base UI — the vaul data-attrs in nova css are
Base UI's vaul-compatible attributes). tw-animate-css kept (globals
imports it). Clean build green; 496 web tests pass.

line-chart-6.tsx deleted (render had zero consumers since the metric
overview rebuild): its Metric* types — the only live exports — moved to
home/components/overview/metric-types.ts, 5 importers repointed.
