+++ apps/web/package.json
@@ -29,25 +29,10 @@
 		"@fontsource/caveat": "^5.2.8",
 		"@headlessui/react": "^2.2.7",
 		"@heroicons/react": "^2.2.0",
-		"@intentui/icons": "^1.11.0",
 		"@lottiefiles/dotlottie-react": "^0.17.8",
 		"@mapbox/search-js-react": "^1.5.1",
 		"@number-flow/react": "^0.5.10",
 		"@onetool/backend": "workspace:*",
-		"@radix-ui/react-avatar": "^1.1.10",
-		"@radix-ui/react-checkbox": "^1.3.3",
-		"@radix-ui/react-collapsible": "^1.1.12",
-		"@radix-ui/react-dialog": "^1.1.15",
-		"@radix-ui/react-dropdown-menu": "^2.1.16",
-		"@radix-ui/react-label": "^2.1.7",
-		"@radix-ui/react-popover": "^1.1.15",
-		"@radix-ui/react-radio-group": "^1.3.8",
-		"@radix-ui/react-scroll-area": "^1.2.10",
-		"@radix-ui/react-select": "^2.2.6",
-		"@radix-ui/react-separator": "^1.1.7",
-		"@radix-ui/react-slot": "^1.2.3",
-		"@radix-ui/react-tabs": "^1.1.13",
-		"@radix-ui/react-tooltip": "^1.2.8",
 		"@react-email/components": "^1.0.1",
 		"@react-email/render": "^2.0.0",
 		"@react-pdf/renderer": "^4.3.0",
@@ -90,9 +75,7 @@
 		"papaparse": "^5.5.3",
 		"pdf-lib": "^1.17.1",
 		"posthog-js": "^1.306.1",
-		"radix-ui": "^1.4.3",
 		"react": "19.2.3",
-		"react-aria-components": "^1.12.1",
 		"react-day-picker": "^9.11.1",
 		"react-dom": "19.2.3",
 		"react-email": "^5.0.8",
@@ -106,9 +89,7 @@
 		"stripe": "22.1.1",
 		"svix": "^1.76.1",
 		"tailwind-merge": "^3.4.0",
-		"tailwind-variants": "^3.1.1",
 		"tunnel-rat": "^0.1.2",
-		"vaul": "^1.1.2",
 		"zod": "^4.1.8"
 	},
 	"devDependencies": {
